+++ helpdesk_stock/models/__init__.py
@@ -1,4 +1,8 @@
 # Copyright (C) 2018 - TODAY, Open Source Integrators
 # License LGPL-3.0 or later (http://www.gnu.org/licenses/lgpl).
 
-from . import helpdesk_ticket, stock
+from . import (
+    helpdesk_ticket,
+    helpdesk_ticket_line,
+    stock
+)
